requirements.txt
# Core
python-dotenv>=1.0
pyyaml>=6.0
pandas>=2.0
duckdb>=0.9

# ML Models
prophet>=1.1.5
torch>=2.0
xgboost>=2.0
scikit-learn>=1.3

# Statistics & Signal Processing
statsmodels>=0.14
scipy>=1.11

# Visualization & Dashboard
streamlit>=1.28
plotly>=5.18

# Data
kaggle>=1.5

# Dev
pytest>=7.0
pytest-cov>=4.0
ruff>=0.1
pre-commit>=3.0
